2dcontext/hit-regions/addHitRegion.nullpathvalidID.html


<!DOCTYPE html>
<html lang="en">
<head>
<meta charset="UTF-8">
<title>addHitRegion throws NotSupportedError when the specified path has no pixels (no path is assigned) with a valid id in fallback content</title>
    <script src="../../resources/testharness.js"></script>
    <script src="../../resources/testharnessreport.js"></script>
    <link rel="author" title="Mark Sadecki">
    <link rel="help" href="http://www.w3.org/TR/2dcontext/#dom-context-2d-addhitregion">
</head>
<body>
<div id="log"></div>
<div><canvas id="canvas"><button id="button">button</button></canvas></div>

<script>
test(function () {
    assert_throws({ name: "NotSupportedError" },
        function () {
            var canvas = document.getElementById("canvas");
            // Reset context
            canvas.width = canvas.width;
            var button = document.getElementById("button");
            var context = canvas.getContext("2d");

            // If the arguments object's id member is null and the arguments object's control member is null,
            // throw a NotSupportedError exception and abort these steps.
            // http://www.w3.org/TR/2dcontext/#dom-context-2d-addhitregion
            context.addHitRegion({id:'button'});
        }, "addHitRegion throws NotSupportedError when the specified path has no pixels (no path is assigned) with a valid id in fallback content.");
}, "addHitRegion throws NotSupportedError when the specified path has no pixels (no path is assigned) with a valid id in fallback content.");
    </script>
    <div id="log"></div>
    <!-- 2.13 addHitRegion is called when the specified path has no pixels (no path is assigned) with a valid id in fallback content. -->
  </body>
</html>
